+++ src/vue/forms/InvestForm.vue
@@ -276,8 +276,8 @@ export default {
       isAccountGeneral: vuexTypes.isAccountGeneral,
       isAccountUsAccredited: vuexTypes.isAccountUsAccredited,
       isAccountUsVerified: vuexTypes.isAccountUsVerified,
-      balances: vuexTypes.accountBalances,
       assets: vuexTypes.assets,
+      accountBalanceByCode: vuexTypes.accountBalanceByCode,
     }),
 
     convertedAmount () {
@@ -292,11 +292,8 @@ export default {
       let quoteAssetBalances = []
 
       this.sale.quoteAssets.forEach(quote => {
-        const balance = this.balances.find(balanceItem => {
-          return balanceItem.asset.code === quote.asset.id
-        })
-
-        if (balance) {
+        const balance = this.accountBalanceByCode(quote.asset.id)
+        if (balance.id) {
           quoteAssetBalances.push(balance)
         }
       })
@@ -402,8 +399,14 @@ export default {
     'form.asset': async function () {
       try {
         this.former.setAttr('investedAssetCode', this.form.asset.code || '')
-        this.former.setAttr('quoteBalanceId', this.balances.find(item => item.asset.code === this.form.asset.code).id || '')
-        this.former.setAttr('saleQuoteAssetPrices', this.sale.quoteAssetPrices[this.form.asset.code] || '')
+        this.former.setAttr(
+          'quoteBalanceId',
+          this.accountBalanceByCode(this.form.asset.code).id || ''
+        )
+        this.former.setAttr(
+          'saleQuoteAssetPrices',
+          this.sale.quoteAssetPrices[this.form.asset.code] || ''
+        )
         if (this.form.asset.code !== this.sale.defaultQuoteAsset) {
           await this.loadAssetPairPrice()
         }
@@ -429,8 +432,7 @@ export default {
       this.former.setAttr('saleId', this.sale.id || '')
       this.former.setAttr('saleBaseAssetCode', this.saleBaseAsset.code || '')
 
-      const baseBalance = this.balances
-        .find(balance => balance.asset.code === this.sale.baseAsset)
+      const baseBalance = this.accountBalanceByCode(this.sale.baseAsset)
 
       if (!baseBalance) {
         const operation = this.former.buildOpCreateBalance()
@@ -438,7 +440,7 @@ export default {
         await this.loadBalances()
       }
 
-      this.former.setAttr('baseBalanceId', this.balances.find(item => item.asset.code === this.sale.baseAsset).id || '')
+      this.former.setAttr('baseBalanceId', baseBalance.id || '')
 
       await this.loadCurrentInvestment()
       this.isLoaded = true