+++ src/vue/forms/TradeForms/YourTradeOfferForm.vue
@@ -1,16 +1,14 @@
 <template>
   <form
     v-if="isLoaded"
-    @submit.prevent="isFormValid() && showConfirmation()"
+    @submit.prevent="tryToSubmit"
   >
     <div class="app__form-row">
       <div class="app__form-field">
-        <div class="app__form-field">
-          <readonly-field
-            :label="baseAssetLabelTranslationId | globalize"
-            :value="offer.baseAsset.id"
-          />
-        </div>
+        <readonly-field
+          :label="baseAssetLabelTranslationId | globalize"
+          :value="former.attrs.baseAssetCode"
+        />
       </div>
     </div>
 
@@ -18,15 +16,15 @@
       <div class="app__form-field">
         <input-field
           v-model.trim="form.price"
-          name="submit-trade-offer-price"
+          name="your-trade-offer-price"
           type="number"
           :min="0"
           :max="config.MAX_AMOUNT"
           :step="config.MIN_AMOUNT"
           :label="
-            'submit-trade-offer-form.price-lbl' | globalize({
-              baseAsset: assetPair.base,
-              quoteAsset: assetPair.quote,
+            'your-trade-offer-form.price-lbl' | globalize({
+              baseAsset: former.attrs.baseAssetCode,
+              quoteAsset: former.attrs.quoteAssetCode,
             })
           "
           :error-message="getFieldErrorMessage(
@@ -46,13 +44,13 @@
       <div class="app__form-field">
         <input-field
           v-model.trim="form.baseAmount"
+          name="your-trade-offer-base-amount"
           type="number"
           :min="0"
           :max="config.MAX_AMOUNT"
           :step="config.MIN_AMOUNT"
-          name="submit-trade-offer-base-amount"
-          :label="'submit-trade-offer-form.base-amount-lbl' | globalize({
-            asset: assetPair.base
+          :label="'your-trade-offer-form.base-amount-lbl' | globalize({
+            asset: former.attrs.baseAssetCode
           })"
           :error-message="getFieldErrorMessage(
             'form.baseAmount',
@@ -62,7 +60,7 @@
               to: config.MAX_AMOUNT,
             }
           )"
-          @blur="touchField('form.baseAmount')"
+          @blur="touchField('form.amount')"
           :disabled="formMixin.isDisabled"
         />
       </div>
@@ -73,18 +71,18 @@
         <template v-if="isFeesLoaded">
           <readonly-field
             :label="
-              'submit-trade-offer-form.total-amount-lbl' | globalize({
-                asset: assetPair.quote
+              'your-trade-offer-form.total-amount-lbl' | globalize({
+                asset: former.attrs.quoteAssetCode
               })
             "
             :value="{
               value: quoteAmount,
-              currency: assetPair.quote,
+              currency: former.attrs.quoteAssetCode,
             } | formatMoney"
             :error-message="getFieldErrorMessage(
               'quoteAmount',
               {
-                available: isBuy ? quoteAssetBalance : baseAssetBalance,
+                available: quoteAssetBalance,
                 from: config.MIN_AMOUNT,
                 to: config.MAX_AMOUNT,
               }
@@ -92,13 +90,13 @@
           />
 
           <fees-renderer
-            class="submit-trade-offer-form__fees"
+            class="your-trade-offer-form__fees"
             :fees-collection="fees"
           />
         </template>
 
         <template v-else>
-          <loader message-id="submit-trade-offer-form.loading-msg" />
+          <loader message-id="your-trade-offer-form.loading-msg" />
         </template>
       </div>
     </div>
@@ -106,7 +104,7 @@
     <template v-if="formMixin.isConfirmationShown">
       <form-confirmation
         class="app__form-confirmation"
-        :is-pending="isOfferSubmitting"
+        :is-pending="isFormSubmitting"
         @ok="submit"
         @cancel="hideConfirmation"
       />
@@ -116,17 +114,22 @@
       <div class="app__form-actions">
         <button
           v-ripple
-          class="app__button-raised submit-trade-offer-form__btn"
           type="submit"
+          @click="setUpdateSubmitMode"
+          class="app__button-raised your-trade-offer-form__btn"
           :disabled="formMixin.isDisabled || !isFeesLoaded"
         >
-          <template v-if="isBuy">
-            {{ 'submit-trade-offer-form.buy-btn' | globalize }}
-          </template>
+          {{ 'your-trade-offer-form.update-order-btn' | globalize }}
+        </button>
 
-          <template v-else>
-            {{ 'submit-trade-offer-form.sell-btn' | globalize }}
-          </template>
+        <button
+          v-ripple
+          type="submit"
+          @click="setCancelSubmitMode"
+          class="app__button-flat your-trade-offer-form__btn"
+          :disabled="formMixin.isDisabled"
+        >
+          {{ 'your-trade-offer-form.cancel-order-btn' | globalize }}
         </button>
       </div>
     </template>
@@ -134,7 +137,7 @@
 
   <loader
     v-else
-    message-id="submit-trade-offer-form.loading-msg" />
+    message-id="your-trade-offer-form.loading-msg" />
 </template>
 
 <script>
@@ -145,16 +148,17 @@ import FeesRenderer from '@/vue/common/fees/FeesRenderer'
 import Loader from '@/vue/common/Loader'
 
 import FormMixin from '@/vue/mixins/form.mixin'
-import OfferManagerMixin from '@/vue/mixins/offer-manager.mixin'
-import FeesMixin from '@/vue/common/fees/fees.mixin'
-
-import { FEE_TYPES } from '@tokend/js-sdk'
 
 import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
+import { TradeFormer } from '@/js/formers/TradeFormer'
+import { vuexTypes } from '@/vuex'
+import { mapGetters, mapActions } from 'vuex'
+import { buildOpCancel } from '@/js/helpers/trade-helper'
 
 import { MathUtil } from '@/js/utils/math.util'
 import config from '@/config'
+import { api } from '@/api'
 
 import {
   required,
@@ -164,13 +168,19 @@ import {
 } from '@validators'
 
 const EVENTS = {
-  offerSubmitted: 'offer-submitted',
+  offerCanceled: 'offer-canceled',
+  offerUpdated: 'offer-updated',
+}
+
+const SUBMIT_MODES = {
+  cancel: 'cancel',
+  update: 'update',
 }
 
 const FEES_LOADING_DELAY_MS = 300
 
 export default {
-  name: 'submit-trade-offer-form',
+  name: 'your-trade-offer-form',
   components: {
     ReadonlyField,
     FeesRenderer,
@@ -178,14 +188,11 @@ export default {
   },
   mixins: [
     FormMixin,
-    OfferManagerMixin,
-    FeesMixin,
   ],
 
   props: {
-    assetPair: { type: Object, required: true },
-    isBuy: { type: Boolean, default: false },
     offer: { type: Object, required: true },
+    former: { type: TradeFormer, default: () => new TradeFormer() },
   },
 
   data: () => ({
@@ -198,7 +205,8 @@ export default {
     isFeesLoaded: false,
     isFeesLoadFailed: false,
     isLoaded: false,
-    isOfferSubmitting: false,
+    submitMode: '',
+    isFormSubmitting: false,
     config,
   }),
 
@@ -213,14 +221,14 @@ export default {
         baseAmount: {
           required,
           decimal,
-          noMoreThanAvailableOnBalance: this.isBuy ||
+          noMoreThanAvailableOnBalance: this.offer.isBuy ||
             maxValueBig(this.baseAssetBalance),
           amountRange: amountRange(config.MIN_AMOUNT, config.MAX_AMOUNT),
         },
       },
 
       quoteAmount: {
-        noMoreThanAvailableOnBalance: !this.isBuy ||
+        noMoreThanAvailableOnBalance: !this.offer.isBuy ||
           maxValueBig(this.quoteAssetBalance),
         amountRange: amountRange(config.MIN_AMOUNT, config.MAX_AMOUNT),
       },
@@ -228,49 +236,66 @@ export default {
   },
 
   computed: {
+    ...mapGetters({
+      accountBalances: vuexTypes.accountBalances,
+      accountId: vuexTypes.accountId,
+      accountBalanceByCode: vuexTypes.accountBalanceByCode,
+    }),
+
     baseAssetLabelTranslationId () {
-      return this.isBuy
-        ? 'submit-trade-offer-form.asset-to-buy-lbl'
-        : 'submit-trade-offer-form.asset-to-sell-lbl'
+      return this.offer.isBuy
+        ? 'your-trade-offer-form.asset-to-buy-lbl'
+        : 'your-trade-offer-form.asset-to-sell-lbl'
+    },
+
+    accountAssets () {
+      return this.accountBalances
+        .map(balance => balance.asset.code)
+        .filter(asset => asset !== this.former.attrs.quoteAssetCode)
     },
 
     baseAssetBalance () {
-      const balanceItem = this.accountBalances
-        .find(balance => balance.asset.code === this.assetPair.base)
-      return balanceItem ? balanceItem.balance : ''
-    },
-
-    quoteAssetBalance () {
-      const balanceItem = this.accountBalances
-        .find(balance => balance.asset.code === this.assetPair.quote)
-      return balanceItem ? balanceItem.balance : ''
-    },
-
-    quoteAmount () {
-      if (this.form.price && this.form.baseAmount) {
-        return MathUtil.multiply(this.form.price, this.form.baseAmount)
+      const balanceItem =
+        this.accountBalanceByCode(this.former.attrs.baseAssetCode)
+      if (balanceItem) {
+        return this.offer.isBuy
+          ? balanceItem.balance
+          : MathUtil.add(balanceItem.balance, this.former.attrs.baseAmount)
       } else {
         return ''
       }
     },
 
-    createOfferOpts () {
-      return {
-        pair: {
-          base: this.assetPair.base,
-          quote: this.assetPair.quote,
-        },
-        baseAmount: this.form.baseAmount,
-        quoteAmount: this.quoteAmount,
-        price: this.form.price,
-        isBuy: this.isBuy,
-        fee: this.fees.totalFee,
+    quoteAssetBalance () {
+      const balanceItem =
+        this.accountBalanceByCode(this.former.attrs.quoteAssetCode)
+
+      if (balanceItem) {
+        return this.offer.isBuy
+          ? balanceItem.balance
+          : MathUtil.add(balanceItem.balance, this.former.attrs.quoteAmount)
+      } else {
+        return ''
+      }
+    },
+
+    quoteAmount () {
+      if (this.form.price && this.form.baseAmount) {
+        let quoteAmount = MathUtil.multiply(
+          this.form.price,
+          this.form.baseAmount
+        )
+        this.former.setAttr('quoteAmount', quoteAmount)
+        return quoteAmount
+      } else {
+        return ''
       }
     },
   },
 
   watch: {
     'form.baseAmount' () {
+      this.former.setAttr('baseAmount', this.form.baseAmount)
       this.tryLoadFees()
     },
 
@@ -281,9 +306,19 @@ export default {
 
   async created () {
     try {
+      this.former.setAttr('creatorAccountId', this.accountId)
+      this.former.setAttr('baseBalanceId',
+        this.accountBalanceByCode(this.former.attrs.baseAssetCode).id)
+      this.former.setAttr('quoteBalanceId',
+        this.accountBalanceByCode(this.former.attrs.quoteAssetCode).id)
+
+      this.form.baseAmount = this.former.attrs.baseAmount
+      this.form.price = String(MathUtil.divide(
+        +this.former.attrs.quoteAmount,
+        +this.former.attrs.baseAmount
+      ))
+
       await this.loadBalances()
-      this.populateForm()
-      await this.loadFees()
       this.isLoaded = true
     } catch (e) {
       ErrorHandler.processWithoutFeedback(e)
@@ -291,10 +326,9 @@ export default {
   },
 
   methods: {
-    populateForm () {
-      this.form.price = this.offer.price
-      this.form.baseAmount = this.offer.baseAmount
-    },
+    ...mapActions({
+      loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
+    }),
 
     tryLoadFees () {
       this.isFeesLoaded = false
@@ -311,13 +345,7 @@ export default {
 
     async loadFees () {
       try {
-        this.fees = await this.calculateFees({
-          assetCode: this.assetPair.quote,
-          amount: this.quoteAmount || 0,
-          senderAccountId: this.accountId,
-          type: FEE_TYPES.offerFee,
-        })
-
+        this.fees = await this.former.calculateFees()
         this.isFeesLoaded = true
       } catch (e) {
         this.isFeesLoadFailed = true
@@ -325,18 +353,45 @@ export default {
       }
     },
 
-    async submit () {
-      this.isOfferSubmitting = true
-      try {
-        await this.createOffer(this.createOfferOpts)
+    setUpdateSubmitMode () {
+      this.submitMode = SUBMIT_MODES.update
+    },
 
-        Bus.success('submit-trade-offer-form.order-submitted-msg')
-        this.$emit(EVENTS.offerSubmitted)
+    setCancelSubmitMode () {
+      this.submitMode = SUBMIT_MODES.cancel
+    },
+
+    async submit () {
+      this.isFormSubmitting = true
+      try {
+        switch (this.submitMode) {
+          case SUBMIT_MODES.cancel:
+            const opCancel = buildOpCancel(this.offer)
+            await api.postOperations(opCancel)
+            Bus.success('your-trade-offer-form.order-canceled-msg')
+            this.$emit(EVENTS.offerCanceled)
+
+            break
+          case SUBMIT_MODES.update:
+            const opCancelOp = buildOpCancel(this.offer)
+            const opCreateOp = await this.former.buildOpCreate()
+            await api.postOperations(opCancelOp, opCreateOp)
+            Bus.success('your-trade-offer-form.order-updated-msg')
+            this.$emit(EVENTS.offerUpdated)
+
+            break
+        }
       } catch (e) {
         ErrorHandler.process(e)
       }
-      this.isOfferSubmitting = false
+
+      this.isFormSubmitting = false
       this.hideConfirmation()
+    },
+
+    tryToSubmit () {
+      if (this.submitMode === SUBMIT_MODES.update && !this.isFormValid()) return
+      this.showConfirmation()
     },
   },
 }
@@ -345,12 +400,12 @@ export default {
 <style lang="scss" scoped>
 @import '../app-form';
 
-.submit-trade-offer-form__btn {
-  max-width: 14rem;
+.your-trade-offer-form__btn {
+  max-width: 16rem;
   width: 100%;
 }
 
-.submit-trade-offer-form__fees {
+.your-trade-offer-form__fees {
   margin-top: 1rem;
 }
 </style>